refactor(WinningNumberSection): 당첨번호와 보너스번호 읽는 로직 컴포넌트에 추가
